pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.6",
    "google-adk>=1.17.0",
    "toolbox-core>=0.5.0",
]
